Create Rubins-rules-t-test
In this file I combined the results from the multiple imputation regression analysis from stata in R. In three steps, first step I created a data frame object in which each column contains the regression results for the imputations m=0 (0 is the results without imputation) till m=40. Each row contains one concept, first row are the beta estimates, second row are the se, third row contains the number of observations and the forth row the r-squared results. I combine the beta, the SE and the R-squared (after z-transforming)  according to Rubins rules with the user defined function results.rubin