Create Graph correlation of RCA for EXGR and value added exports with GDP per capita (2005 const. US Dollar)
Here I create the data frames for the plot the (rank and tau) correlation of the Ricardo comparative advantage measure for EXGR and different value added export measures with GDP per capita. The graph is a scatter plot with a linear best fit in the graph and the regression eq. estimates in it.